requirements.txt
pandas>=1.5
python-dotenv>=1.0.0

streamlit>=1.32.0

openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.5

typing-extensions>=4.5
langgraph>=0.0.30

